- Adds DigitalForm class as a wrapper around DigitalForm nodes. - Moves some logic from DigitalFormsService into the DigitalForm wrapper class to be more OOP.
